+++ OrbitCaptureClient/include/OrbitCaptureClient/CaptureClient.h
@@ -11,8 +11,8 @@
 #include "OrbitBase/ThreadPool.h"
 #include "OrbitClientData/ModuleManager.h"
 #include "OrbitClientData/ProcessData.h"
-#include "TracepointCustom.h"
-#include "UserDefinedCaptureData.h"
+#include "OrbitClientData/TracepointCustom.h"
+#include "OrbitClientData/UserDefinedCaptureData.h"
 #include "absl/container/flat_hash_set.h"
 #include "capture_data.pb.h"
 #include "grpcpp/channel.h"
