+++ OrbitClientData/TracepointInfoManagerTest.cpp
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-#include "TracepointInfoManager.h"
+#include "OrbitClientData/TracepointInfoManager.h"
 #include "gtest/gtest.h"
 
 TEST(TracepointInfoManager, Contains) {
