+++ OrbitClientData/UserDefinedCaptureData.cpp
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-#include "UserDefinedCaptureData.h"
+#include "OrbitClientData/UserDefinedCaptureData.h"
 
 #include "OrbitClientData/FunctionUtils.h"
