+++ OrbitClientGgp/ClientGgp.cpp
@@ -15,14 +15,13 @@
 #include "OrbitBase/Result.h"
 #include "OrbitCaptureClient/CaptureClient.h"
 #include "OrbitClientData/FunctionUtils.h"
-#include "OrbitClientData/ModuleData.h"
 #include "OrbitClientData/ProcessData.h"
+#include "OrbitClientData/UserDefinedCaptureData.h"
 #include "OrbitClientModel/CaptureSerializer.h"
+#include "OrbitClientModel/SamplingProfiler.h"
 #include "OrbitClientServices/ProcessManager.h"
-#include "SamplingProfiler.h"
 #include "StringManager.h"
 #include "SymbolHelper.h"
-#include "UserDefinedCaptureData.h"
 #include "absl/container/flat_hash_map.h"
 #include "absl/container/flat_hash_set.h"
 #include "capture_data.pb.h"
