+++ OrbitClientModel/CaptureDeserializerTest.cpp
@@ -9,8 +9,8 @@
 #include "OrbitClientData/ModuleData.h"
 #include "OrbitClientData/ModuleManager.h"
 #include "OrbitClientData/ProcessData.h"
+#include "OrbitClientData/UserDefinedCaptureData.h"
 #include "OrbitClientModel/CaptureDeserializer.h"
-#include "UserDefinedCaptureData.h"
 #include "absl/base/casts.h"
 #include "capture_data.pb.h"
 #include "gmock/gmock-actions.h"
