+++ OrbitGl/App.cpp
@@ -19,7 +19,6 @@
 #include <utility>
 
 #include "CallStackDataView.h"
-#include "Callstack.h"
 #include "CaptureWindow.h"
 #include "CoreUtils.h"
 #include "Disassembler.h"
@@ -32,16 +31,17 @@
 #include "OrbitBase/Logging.h"
 #include "OrbitBase/Result.h"
 #include "OrbitBase/Tracing.h"
+#include "OrbitClientData/Callstack.h"
 #include "OrbitClientData/FunctionUtils.h"
 #include "OrbitClientData/ModuleData.h"
 #include "OrbitClientData/ModuleManager.h"
 #include "OrbitClientData/ProcessData.h"
 #include "OrbitClientModel/CaptureDeserializer.h"
 #include "OrbitClientModel/CaptureSerializer.h"
+#include "OrbitClientModel/SamplingProfiler.h"
 #include "Path.h"
 #include "PresetsDataView.h"
 #include "ProcessesDataView.h"
-#include "SamplingProfiler.h"
 #include "SamplingReport.h"
 #include "StringManager.h"
 #include "SymbolHelper.h"
