+++ OrbitGl/App.h
@@ -16,8 +16,6 @@
 #include "ApplicationOptions.h"
 #include "CallStackDataView.h"
 #include "CallTreeView.h"
-#include "Callstack.h"
-#include "CallstackData.h"
 #include "CaptureWindow.h"
 #include "DataManager.h"
 #include "DataView.h"
@@ -36,10 +34,13 @@
 #include "OrbitBase/ThreadPool.h"
 #include "OrbitCaptureClient/CaptureClient.h"
 #include "OrbitCaptureClient/CaptureListener.h"
+#include "OrbitClientData/Callstack.h"
+#include "OrbitClientData/CallstackData.h"
 #include "OrbitClientData/FunctionInfoSet.h"
 #include "OrbitClientData/ModuleData.h"
 #include "OrbitClientData/ModuleManager.h"
 #include "OrbitClientData/ProcessData.h"
+#include "OrbitClientData/TracepointCustom.h"
 #include "OrbitClientServices/CrashManager.h"
 #include "OrbitClientServices/ProcessManager.h"
 #include "OrbitClientServices/TracepointServiceClient.h"
@@ -53,7 +54,6 @@
 #include "StringManager.h"
 #include "SymbolHelper.h"
 #include "Threading.h"
-#include "TracepointCustom.h"
 #include "TracepointsDataView.h"
 #include "absl/container/flat_hash_map.h"
 #include "absl/container/flat_hash_set.h"
