+++ OrbitGl/CallStackDataView.cpp
@@ -5,7 +5,7 @@
 #include "CallStackDataView.h"
 
 #include "App.h"
-#include "Callstack.h"
+#include "OrbitClientData/Callstack.h"
 #include "OrbitClientData/FunctionUtils.h"
 #include "Path.h"
 #include "absl/flags/flag.h"
