+++ OrbitGl/TimeGraph.cpp
@@ -19,8 +19,8 @@
 #include "GraphTrack.h"
 #include "ManualInstrumentationManager.h"
 #include "OrbitClientData/FunctionUtils.h"
+#include "OrbitClientModel/SamplingProfiler.h"
 #include "PickingManager.h"
-#include "SamplingProfiler.h"
 #include "SchedulerTrack.h"
 #include "StringManager.h"
 #include "TextBox.h"
