+++ OrbitGl/TimeGraph.h
@@ -11,7 +11,6 @@
 #include "AsyncTrack.h"
 #include "Batcher.h"
 #include "BlockChain.h"
-#include "CaptureData.h"
 #include "FrameTrack.h"
 #include "Geometry.h"
 #include "GpuTrack.h"
@@ -19,6 +18,7 @@
 #include "ManualInstrumentationManager.h"
 #include "OrbitBase/Profiling.h"
 #include "OrbitBase/Tracing.h"
+#include "OrbitClientModel/CaptureData.h"
 #include "SchedulerTrack.h"
 #include "StringManager.h"
 #include "TextBox.h"
